+++ packages/desktop_drop/lib/desktop_drop.dart
@@ -1,2 +1,1 @@
-export 'src/events.dart';
 export 'src/drop_target.dart';