+++ packages/desktop_drop/lib/src/drop_target.dart
@@ -5,7 +5,33 @@ import 'package:flutter/widgets.dart';
 import 'channel.dart';
 import 'events.dart';
 
-typedef OnDragDoneCallback = void Function(List<Uri> urls);
+@immutable
+class DropDoneDetails {
+  const DropDoneDetails({
+    required this.urls,
+    required this.localPosition,
+    required this.globalPosition,
+  });
+
+  final List<Uri> urls;
+  final Offset localPosition;
+  final Offset globalPosition;
+}
+
+class DropEventDetails {
+  DropEventDetails({
+    required this.localPosition,
+    required this.globalPosition,
+  });
+
+  final Offset localPosition;
+
+  final Offset globalPosition;
+}
+
+typedef OnDragDoneCallback = void Function(DropDoneDetails details);
+
+typedef OnDragCallback<Detail> = void Function(Detail details);
 
 class DropTarget extends StatefulWidget {
   const DropTarget({
@@ -20,10 +46,14 @@ class DropTarget extends StatefulWidget {
 
   final Widget child;
 
-  final VoidCallback? onDragEntered;
-  final VoidCallback? onDragExited;
+  /// Callback when drag entered target area.
+  final OnDragCallback<DropEventDetails>? onDragEntered;
 
-  final VoidCallback? onDragUpdated;
+  /// Callback when drag exited target area.
+  final OnDragCallback<DropEventDetails>? onDragExited;
+
+  /// Callback when drag hover on target area.
+  final OnDragCallback<DropEventDetails>? onDragUpdated;
 
   final OnDragDoneCallback? onDragDone;
 
@@ -59,7 +89,11 @@ class _DropTargetState extends State<DropTarget> {
     } else if (!widget.enable && oldWidget.enable) {
       DesktopDrop.instance.removeRawDropEventListener(_onDropEvent);
       if (_status != _DragTargetStatus.idle) {
-        _updateStatus(_DragTargetStatus.idle);
+        _updateStatus(
+          _DragTargetStatus.idle,
+          localLocation: Offset.zero,
+          globalLocation: Offset.zero,
+        );
       }
     }
   }
@@ -69,45 +103,86 @@ class _DropTargetState extends State<DropTarget> {
     if (renderBox == null) {
       return;
     }
-    final position =
-        renderBox.globalToLocal(_scaleHoverPoint(context, event.location));
+    final globalPosition = _scaleHoverPoint(context, event.location);
+    final position = renderBox.globalToLocal(globalPosition);
     bool inBounds = renderBox.paintBounds.contains(position);
     if (event is DropEnterEvent) {
       if (!inBounds) {
         assert(_status == _DragTargetStatus.idle);
       } else {
-        _updateStatus(_DragTargetStatus.enter);
+        _updateStatus(
+          _DragTargetStatus.enter,
+          globalLocation: globalPosition,
+          localLocation: position,
+        );
       }
     } else if (event is DropUpdateEvent) {
       if (_status == _DragTargetStatus.idle && inBounds) {
-        _updateStatus(_DragTargetStatus.enter);
-      } else if (_status == _DragTargetStatus.enter && inBounds) {
-        _updateStatus(_DragTargetStatus.update);
+        _updateStatus(
+          _DragTargetStatus.enter,
+          globalLocation: globalPosition,
+          localLocation: position,
+        );
+      } else if ((_status == _DragTargetStatus.enter ||
+              _status == _DragTargetStatus.update) &&
+          inBounds) {
+        _updateStatus(
+          _DragTargetStatus.update,
+          globalLocation: globalPosition,
+          localLocation: position,
+          debugRequiredStatus: false,
+        );
       } else if (_status != _DragTargetStatus.idle && !inBounds) {
-        _updateStatus(_DragTargetStatus.idle);
+        _updateStatus(
+          _DragTargetStatus.idle,
+          globalLocation: globalPosition,
+          localLocation: position,
+        );
       }
     } else if (event is DropExitEvent && _status != _DragTargetStatus.idle) {
-      _updateStatus(_DragTargetStatus.idle);
+      _updateStatus(
+        _DragTargetStatus.idle,
+        globalLocation: globalPosition,
+        localLocation: position,
+      );
     } else if (event is DropDoneEvent &&
         (_status != _DragTargetStatus.idle || Platform.isLinux) &&
         inBounds) {
-      _updateStatus(_DragTargetStatus.idle, required: false);
-      widget.onDragDone?.call(event.uris);
+      _updateStatus(
+        _DragTargetStatus.idle,
+        debugRequiredStatus: false,
+        globalLocation: globalPosition,
+        localLocation: position,
+      );
+      widget.onDragDone?.call(DropDoneDetails(
+        urls: event.uris,
+        localPosition: position,
+        globalPosition: globalPosition,
+      ));
     }
   }
 
-  void _updateStatus(_DragTargetStatus status, {bool required = true}) {
-    assert(!required || _status != status);
+  void _updateStatus(
+    _DragTargetStatus status, {
+    bool debugRequiredStatus = true,
+    required Offset localLocation,
+    required Offset globalLocation,
+  }) {
+    assert(!debugRequiredStatus || _status != status);
     _status = status;
+    final details = DropEventDetails(
+      localPosition: localLocation,
+      globalPosition: globalLocation,
+    );
     switch (_status) {
       case _DragTargetStatus.enter:
-        widget.onDragEntered?.call();
+        widget.onDragEntered?.call(details);
         break;
       case _DragTargetStatus.update:
-        widget.onDragUpdated?.call();
+        widget.onDragUpdated?.call(details);
         break;
       case _DragTargetStatus.idle:
-        widget.onDragExited?.call();
+        widget.onDragExited?.call(details);
         break;
     }
   }